renamed combobox files and split filters.tsx
